Move validator contracts
