+++ packages/contracts/source/contracts/consensus/consensus.ts
@@ -1,9 +1,7 @@
 import { IProcessableUnit } from "../block-processor";
 import {
-	IBlock,
 	IBlockCommit,
 	ICommittedBlock,
-	IKeyPair,
 	IPrecommit,
 	IPrevote,
 	IProposal,
@@ -110,21 +108,6 @@ export interface IScheduler {
 export interface IProposerPicker {
 	handleCommittedBlock(block: IBlockCommit): void;
 	getValidatorIndex(round: number): number;
-}
-
-export interface IValidator {
-	configure(publicKey: string, keyPair: IKeyPair): IValidator;
-	getWalletPublicKey(): string;
-	getConsensusPublicKey(): string;
-	prepareBlock(height: number, round: number): Promise<IBlock>;
-	propose(round: number, block: IBlock, lockProof?: IProposalLockProof): Promise<IProposal>;
-	prevote(height: number, round: number, blockId: string | undefined): Promise<IPrevote>;
-	precommit(height: number, round: number, blockId: string | undefined): Promise<IPrecommit>;
-}
-
-export interface IValidatorRepository {
-	getValidator(publicKey: string): IValidator | undefined;
-	getValidators(publicKeys: string[]): IValidator[];
 }
 
 export interface IProcessor {
