+++ packages/contracts/source/contracts/index.ts
@@ -13,4 +13,5 @@ export * as State from "./state";
 export * as TransactionPool from "./transaction-pool";
 export * as Transactions from "./transactions";
 export * as Types from "./types";
+export * as Validator from "./validator";
 export * as ValidatorSet from "./validator-set";
